move sqs code
