+++ src/sqs.py
@@ -1,0 +1,33 @@
+#!/bin/env python3
+"""
+sqs:
+- helper to work with sqs
+"""
+
+import boto3
+import json
+import os
+from time import time
+from src import logger
+
+
+sqs = boto3.resource('sqs')
+SQS_QUEUE_NAME = os.getenv('SQS_QUEUE_NAME', 'committee-sync-queue')
+queue = sqs.get_queue_by_name(QueueName=SQS_QUEUE_NAME)
+
+def pull_message_from_sqs() -> Dict[str, Any]:
+    """pulls a committee id from SQS
+    see https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sqs.html#SQS.Client.receive_message
+
+    Returns:
+        Any: SQS Message object
+    """
+
+    start_time = time()
+    time_to_end = start_time + 10
+    while time() < time_to_end:
+        message = queue.receive_messages(MaxNumberOfMessages=1)
+        if message:
+            return message[0]
+    logger.info('No messages received, exiting')
+    return {}
